rename common-validator to shared-validator
