+++ src/validator/user.validator.ts
@@ -1,5 +1,5 @@
 import {checkSchema} from "express-validator";
-import {commonStringRule, passwordRule, emailRule, personNameRule} from "./common.validator";
+import {commonStringRule, passwordRule, emailRule, personNameRule} from "./shared.validator";
 import {userCheckEmailValidator} from "./custom/validator";
 
 export const userSignUpVS = checkSchema({
